* Added ParseBuffer for shuffling parsing objects to wasm. * Added fast float optimised returns for parsing arrays of numbers/vector types/triangles. * Added special case geometric tolerance padding to subtracted   geometry objects in CSG that are not themselves to be subtracted. * Made it so no parsed points circled through floats and reduced precision prematurely. * Added box overlap tests to early out inside/outside GWN testing. * Switched over to using a non-predicate determinant for GWN calculations. * Added fast rescale relative point option for geometry. * Fixed asserts for prefix sum map. * Added preliminary ray support. * Added CSG memoization for performance (can be disabled for low-memory mode).

+++ conway_geometry/operations/curve_utils.h
@@ -408,7 +408,7 @@ inline bool MatrixFlipsTriangles(const glm::dmat3 &mat) {
 }
 
 //TODO: support IfcAlignment types  
-inline IfcCurve GetEllipseCurve(float radiusX, float radiusY, int numSegments,
+inline IfcCurve GetEllipseCurve(double radiusX, double radiusY, int numSegments,
                                 glm::dmat3 placement = glm::dmat3(1),
                                 double startRad = 0,
                                 double endRad = M_PI * 2,
@@ -507,7 +507,7 @@ inline IfcCurve GetEllipseCurve(float radiusX, float radiusY, int numSegments,
   return c;
 }
 
-inline IfcCurve GetCircleCurve(float radius, int numSegments,
+inline IfcCurve GetCircleCurve(double radius, int numSegments,
                                glm::dmat3 placement = glm::dmat3(1)) {
   return GetEllipseCurve(radius, radius, numSegments, placement);
 }
